requirements.txt
# Core utilities
python-dotenv==1.2.1
requests==2.32.5

# Translation
deep-translator==1.11.4

# Image processing
Pillow==12.1.1

# Web App
streamlit==1.54.0